refactor: updating useMeasure hook
